Refactor NODX playground examples and enhance rendering features
- Updated app.js to improve sample loading and rendering logic, including new render options for documentation layout.
- Introduced renderDocsPreview and related functions to support a documentation shell layout.
- Enhanced error handling and diagnostics in the rendering process.
- Added new playground examples for inline formatting, variables, shared styles, custom theme tokens, and documentation themes.
- Created a README for the playground examples directory to clarify usage.
- Updated styles.css to accommodate new documentation layout styles.
- Improved the inflate.rs logic for better handling of Huffman blocks.
- Added tests for playground examples and rendering HTML to ensure validity and correctness.
- Updated build_site.py to mirror static directories for the playground and JavaScript package files.
- Modified index.html to include a link to the playground for easier access.

+++ crates/nodx-package/src/inflate.rs
@@ -153,15 +153,14 @@ fn inflate_huffman_block(
             return Ok(());
         } else if symbol <= 285 {
             let idx = symbol as usize - 257;
-            let length =
-                LENGTH_BASE[idx] as usize + reader.read_bits(LENGTH_EXTRA[idx] as u8)? as usize;
+            let length = LENGTH_BASE[idx] as usize + reader.read_bits(LENGTH_EXTRA[idx])? as usize;
             let dist_symbol = decode_symbol(reader, dist)?;
             if dist_symbol >= 30 {
                 return Err("invalid distance symbol");
             }
             let didx = dist_symbol as usize;
             let distance = DISTANCE_BASE[didx] as usize
-                + reader.read_bits(DISTANCE_EXTRA[didx] as u8)? as usize;
+                + reader.read_bits(DISTANCE_EXTRA[didx])? as usize;
             if distance == 0 || distance > out.len() {
                 return Err("invalid back reference");
             }
